Refactor methods into functions.
